feat: inclusão unit test integration - srv delete

+++ java-spring-expert/dscatalog/backend/src/main/java/com/devsuperior/dscatalog/services/ProductService.java
@@ -88,14 +88,17 @@ public class ProductService {
         }
     }
 
-    @Transactional(propagation = Propagation.SUPPORTS)
+    @Transactional(propagation = Propagation.REQUIRED)
     public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Id not found " + id);
+        }
         try {
             repository.deleteById(id);
         }
-        catch (EmptyResultDataAccessException e) {
-            throw new ResourceNotFoundException("Id not found " + id);
-        }
+        //catch (EmptyResultDataAccessException e) {
+        //    throw new ResourceNotFoundException("Id not found " + id);
+        //}
         catch (DataIntegrityViolationException e) {
             throw new DatabaseException("Falha de integridade referencial");
         }
